Rename proxy to server
