+++ proxy/main.go
@@ -5,7 +5,7 @@ import (
 	"log"
 	"strconv"
 	"path/filepath"
-	"github.com/ktk-2005/Feedbacker-Forum/proxy/proxy"
+	"github.com/ktk-2005/Feedbacker-Forum/proxy/server"
 	"github.com/ktk-2005/Feedbacker-Forum/proxy/resolver"
 )
 
@@ -30,22 +30,22 @@ func main() {
 	path := filepath.Dir(os.Args[0])
 
 	var resolverConfig resolver.Config
-	var proxyConfig proxy.Config
+	var serverConfig server.Config
 
 	resolverConfig.DbDriver = getEnv("FFGP_DB_DRIVER", "sqlite3")
-	resolverConfig.DbConnectString = getEnv("FFGP_DB_DRIVER", filepath.Join(path, "../server/dev_db.sqlite"))
-	proxyConfig.ProxyPort = getEnvInt("FFGP_PROXY_PORT", "8086")
-	proxyConfig.ErrorPort = getEnvInt("FFGP_ERROR_PORT", "8087")
-	proxyConfig.InjectScript = getEnv("FFGP_INJECT_SCRIPT", "http://localhost:8080/main.js")
+	resolverConfig.DbConnectString = getEnv("FFGP_DB_CONNECT", filepath.Join(path, "../server/dev_db.sqlite"))
+	serverConfig.ProxyPort = getEnvInt("FFGP_PROXY_PORT", "8086")
+	serverConfig.ErrorPort = getEnvInt("FFGP_ERROR_PORT", "8087")
+	serverConfig.InjectScript = getEnv("FFGP_INJECT_SCRIPT", "http://localhost:8080/main.js")
 
 	err := resolver.Initialize(&resolverConfig)
 	if err != nil {
 		log.Fatalf("Failed to initialize container database: %v", err)
 	}
 
-	err = proxy.StartServing(&proxyConfig)
+	err = server.StartServing(&serverConfig)
 	if err != nil {
-		log.Fatalf("Failed to start proxy: %v", err)
+		log.Fatalf("Failed to start server: %v", err)
 	}
 }
