+++ proxy/server/server.go
@@ -1,4 +1,4 @@
-package proxy
+package server
 
 import (
 	"bytes"
